# Conflicts:
#	src/main/AndroidManifest.xml
#	src/main/java/com/nmc/android/ui/SplashActivity.kt
#	src/main/res/values-b+en+001/strings.xml
#	src/main/res/values/dimens.xml
#	src/main/res/values/strings.xml

+++ src/main/java/com/nextcloud/client/media/Player.kt
@@ -61,7 +61,7 @@ internal class Player(
     private var enqueuedFile: PlaylistItem? = null
 
     private var playedFile: OCFile? = null
-    private var startPositionMs: Int = 0
+    private var startPositionMs: Long = 0
     private var autoPlay = true
     private var user: User? = null
     private var dataSource: String? = null
